+++ {{cookiecutter.repo_name}}/python/{{cookiecutter.package_name|lower}}/__init__.py
@@ -8,8 +8,8 @@ import os
 import yaml
 
 # Inits the logging system. Only shell logging, and exception and warning catching.
-# File logging can be started by calling log.start_file_logger(name).
-from .utils import log
+# File logging can be started by calling log.start_file_logger(path).
+from .utils import get_logger
 
 
 def merge(user, default):
@@ -30,12 +30,11 @@ NAME = '{{cookiecutter.package_name}}'
 
 # Loads config
 yaml_kwds = dict()
-if parse_version(yaml.__version__) >= parse_version("5.1"):
+if parse_version(yaml.__version__) >= parse_version('5.1'):
     yaml_kwds.update(Loader=yaml.FullLoader)
 
-
-config_path = os.path.join(os.path.dirname(__file__), "/etc/{0}.yml".format(NAME))
-with open(config_path, "r") as fp:
+config_path = os.path.join(os.path.dirname(__file__), 'etc/{0}.yml'.format(NAME))
+with open(config_path, 'r') as fp:
     config = yaml.load(fp, **yaml_kwds)
 
 # If there is a custom configuration file, updates the defaults using it.
@@ -43,5 +42,9 @@ custom_config_path = os.path.expanduser('~/.{0}/{0}.yml'.format(NAME))
 if os.path.exists(custom_config_path):
     with open(custom_config_path, "r") as fp:
         config = merge(yaml.load(fp, **yaml_kwds), config)
+
+
+log = get_logger(NAME)
+
 
 __version__ = '{{cookiecutter.version}}'