+++ {{cookiecutter.repo_name}}/python/{{cookiecutter.package_name|lower}}/utils/__init__.py
@@ -1,2 +1,2 @@
 
-from .logger import log
+from .logger import *